% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch-package.R
\docType{package}
\name{arch-package}
\alias{arch}
\alias{arch-package}
\title{arch: 'Arrow' R and C Helpers}
\description{
The 'Apache Arrow' project offers a zero-dependency C interface which this package uses to create and proces 'Arrow' objects -- without requiring the 'arrow' package. This package builds upon a lot of earlier work in package 'narrow' (formerly also named 'sparrow', 'carrow', 'arrowc') by Dewey.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eddelbuettel/arch}
  \item \url{https://github.com/paleolimbot/narrow}
  \item Report bugs at \url{https://github.com/eddelbuettel/arch/issues}
}

}
\keyword{internal}
