% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arch_array}
\alias{arch_array}
\alias{arch_array_validate}
\alias{as_arch_array}
\alias{as_arch_array.arch_array}
\title{Create arrow vectors}
\usage{
arch_array(
  schema = arch_schema("n"),
  array_data = arch_array_data(),
  validate = TRUE
)

arch_array_validate(x)

as_arch_array(x, ...)

\method{as_arch_array}{arch_array}(x, ...)
}
\arguments{
\item{schema}{An \code{\link[=arch_schema]{arch_schema()}}}

\item{array_data}{An \code{\link[=arch_array_data]{arch_array_data()}}}

\item{validate}{Use \code{FALSE} to skip validation}

\item{x}{An object to convert to an \code{\link[=arch_array]{arch_array()}}}

\item{...}{Passed to S3 methods}
}
\value{
An object of class "arch_array"
}
\description{
Create arrow vectors
}
\examples{
arch_array()

}
