% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{arch_deep_copy}
\alias{arch_deep_copy}
\title{Create copies of Arrow vectors}
\usage{
arch_deep_copy(x)
}
\arguments{
\item{x}{An \code{\link[=arch_array]{arch_array()}}.}
}
\value{
An \code{\link[=arch_array]{arch_array()}}
}
\description{
Create copies of Arrow vectors
}
\examples{
arch_deep_copy(as_arch_array(1:5))

}
