% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-ptype.R
\name{arch_default_ptype}
\alias{arch_default_ptype}
\title{Guess prototypes for Arrow schemas}
\usage{
arch_default_ptype(schema)
}
\arguments{
\item{schema}{An \code{\link[=arch_schema]{arch_schema()}}}
}
\value{
A zero-length array prototype
}
\description{
Guess prototypes for Arrow schemas
}
\examples{
arch_default_ptype(arch_schema("n"))
arch_default_ptype(
  arch_schema("+s", children = list(arch_schema("i", "colname")))
)

}
