% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointers.R
\name{arch_pointer_is_valid}
\alias{arch_pointer_is_valid}
\alias{arch_pointer_release}
\alias{arch_pointer_move}
\alias{arch_pointer_export}
\alias{arch_allocate_schema}
\alias{arch_allocate_array_data}
\alias{arch_allocate_array_stream}
\alias{arch_pointer_addr_chr}
\alias{arch_schema_from_pointer}
\alias{arch_array_data_from_pointer}
\alias{arch_array_stream_from_pointer}
\alias{arch_from_pointer}
\title{Danger zone: low-level pointer operations}
\usage{
arch_pointer_is_valid(ptr)

arch_pointer_release(ptr)

arch_pointer_move(ptr_src, ptr_dst)

arch_pointer_export(ptr_src, ptr_dst)

arch_allocate_schema()

arch_allocate_array_data()

arch_allocate_array_stream()

arch_pointer_addr_chr(ptr)

arch_schema_from_pointer(ptr)

arch_array_data_from_pointer(ptr)

arch_array_stream_from_pointer(ptr)

arch_from_pointer(ptr, cls)
}
\arguments{
\item{ptr, ptr_src, ptr_dst}{An external pointer to a \code{\link[=arch_schema]{arch_schema()}},
a \code{\link[=arch_array_data]{arch_array_data()}}, or a \code{\link[=arch_array_stream]{arch_array_stream()}}.}

\item{cls}{One of "arch_schema", "arch_array_data", or
"arch_array_stream".}
}
\description{
Danger zone: low-level pointer operations
}
