% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{arch_schema}
\alias{arch_schema}
\alias{arch_schema_flags}
\alias{arch_schema_deep_copy}
\alias{as_arch_schema}
\alias{as_arch_schema.arch_schema}
\alias{as_arch_schema.character}
\alias{arch_schema_info}
\title{Create Arrow Schema objects}
\usage{
arch_schema(
  format,
  name = "",
  metadata = NULL,
  flags = arch_schema_flags(),
  children = NULL,
  dictionary = NULL
)

arch_schema_flags(
  dictionary_ordered = FALSE,
  nullable = TRUE,
  map_keys_sorted = FALSE
)

arch_schema_deep_copy(schema)

as_arch_schema(x, ...)

\method{as_arch_schema}{arch_schema}(x, ...)

\method{as_arch_schema}{character}(x, ...)

arch_schema_info(x, ..., recursive = FALSE)
}
\arguments{
\item{format}{A
\href{https://arrow.apache.org/docs/format/CDataInterface.html#data-type-description-format-strings}{format string}.}

\item{name}{An optional name}

\item{metadata}{A named character vector of \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}}
containing key/value metadata.}

\item{flags}{Flags to set on input (see \code{\link[=arch_schema_flags]{arch_schema_flags()}})}

\item{children}{A \code{\link[=list]{list()}} of objects created using \code{\link[=arch_schema]{arch_schema()}}.}

\item{dictionary}{An \code{\link[=arch_schema]{arch_schema()}} if this is a dictionary type.}

\item{dictionary_ordered}{\code{TRUE} if the \code{dictionary} array is ordered}

\item{nullable}{\code{TRUE} if the type is semantically nullable}

\item{map_keys_sorted}{\code{TRUE} if the keys for a map have been sorted}

\item{schema}{A schema created with \code{\link[=arch_schema]{arch_schema()}}}

\item{x}{An object to convert to an \code{\link[=arch_schema]{arch_schema()}}}

\item{...}{Passed to S3 methods}

\item{recursive}{Use \code{TRUE} to serialize the schema recursively}
}
\value{
An external pointer with class 'arch_schema'
}
\description{
Create Arrow Schema objects
}
