% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{from_arch_array.Array}
\alias{from_arch_array.Array}
\alias{from_arch_array.R6ClassGenerator}
\alias{as_arch_schema.DataType}
\alias{as_arch_schema.Field}
\alias{as_arch_schema.Schema}
\alias{as_arch_array.Scalar}
\alias{as_arch_array.Array}
\alias{as_arch_array.RecordBatch}
\alias{arch_array_stream_to_arrow}
\alias{as_arch_array_stream.RecordBatchReader}
\alias{as_arch_array_stream.Scanner}
\alias{as_arch_array_stream.Dataset}
\alias{as_arch_array_stream.Table}
\alias{as_arch_array_stream.RecordBatchFileReader}
\alias{as_arch_array_stream.RecordBatchStreamReader}
\alias{as_arch_array_stream.Array}
\alias{as_arch_array_stream.ChunkedArray}
\title{Convert to and from arrow package types}
\usage{
\method{from_arch_array}{Array}(x, ptype, ...)

\method{from_arch_array}{R6ClassGenerator}(x, ptype, ...)

\method{as_arch_schema}{DataType}(x, ...)

\method{as_arch_schema}{Field}(x, ...)

\method{as_arch_schema}{Schema}(x, ...)

\method{as_arch_array}{Scalar}(x, ...)

\method{as_arch_array}{Array}(x, ...)

\method{as_arch_array}{RecordBatch}(x, ...)

arch_array_stream_to_arrow(x)

\method{as_arch_array_stream}{RecordBatchReader}(x, ...)

\method{as_arch_array_stream}{Scanner}(x, ...)

\method{as_arch_array_stream}{Dataset}(x, ...)

\method{as_arch_array_stream}{Table}(x, ...)

\method{as_arch_array_stream}{RecordBatchFileReader}(x, ...)

\method{as_arch_array_stream}{RecordBatchStreamReader}(x, ...)

\method{as_arch_array_stream}{Array}(x, ...)

\method{as_arch_array_stream}{ChunkedArray}(x, ...)
}
\arguments{
\item{x}{An object to convert to or from 'arrow' package types}

\item{ptype}{An R object to use as a prototype}

\item{...}{Passed to S3 methods}
}
\description{
Convert to and from arrow package types
}
